Provide synchronous counterparts for async functions (local and remote) to allow for naturally synchronous calls on the client-side instead of using the dangerous 'sync-over-async' anti-pattern. 'Sync-over-async' is still used on the server-side and in unit tests See http://blog.stephencleary.com/2012/07/dont-block-on-async-code.html
A half-way elegant WCF technique was used here. We re-declare IGenericServiceDispatcherEx interface on the client-side which copies all methods of the original IGenericServiceDispatcher server-side interface. It extends the latter with synchronous operations for matching async counterparts. Unfortunately we can not inherit IGenericServiceDispatcher because FaultContract attributes must be assigned slightly differently (very questionable requirement of WCF). As a result we are able to call certain operations synchronously on the client, while keeping the serve-side implementation asynchronous (for scalability).