Skip tests on client-side execution of query parts 'Client_Join_select_many' and 'Where_subquery_on_navigation_client_eval'
We have to confess the entire topic of client-side evaluation becomes difficult with our chosen approach:
- EFC's RequiresClientXXX properties and logic reside in 'Microsoft.EntityFrameworkCore.Relational' assembly which we practically don't use (and should not use for many other reasons). Re-implementing it doesn't sound very attractive right now.
- Client-side execution partially works in some natural cases, e.g. all tests on 'Northwind.Customer.IsLondon' pass, because the property is directly available on the server side so the actual relational provider is able to make the necessary use of it.
- Tests like 'Where_bool_client_side_negated' calling static 'ClientFunc' method also pass in the current setup, but strictly speaking in the real life scenarios clients should not rely on the presence of all involved methods on the server end.
- The remaining tests 'Client_Join_select_many' and 'Where_subquery_on_navigation_client_eval' have been suppressed because they are calling non-static methods of objects that are not available (and not constructible) on the server side.

Eventually we should implement a stricter mechanism which won't relay execution of client functions to the server, not only because they are no always available, but mainly because we DON'T WANT them to run on the server end. Imagine a client query like ctx.Customers.SelectMany(c => Directory.GetFiles(@"C:\WINDOWS\")). It would be dangerous to transparently execute it on the server, wouldn't it? If we ever get into this topic, I would suggest a restriction to only functions from Enumerable and Queryable classes and functions/properties directly available in entity classes, everything else should never be attempted to run on the server side.

For now I would simply recommend clients to refrain for calling arbitrary functions/properties in the queries. Even if the query succeeds the results may be quite unexpected.